+++ src/components/ProjectView/GithubTab/NpmCard/Dependencies.js
@@ -1,55 +1,39 @@
 import React from 'react'
+
 import Toggle from 'react-toggled'
 import { Link } from 'react-router-dom'
+import styled from 'styled-components'
 
-import withPackageData from '../../../containers/withPackageData'
-import StarTotal from '../../common/utils/StarTotal'
+import withPackageData from '../../../../containers/withPackageData'
+import StarTotal from '../../../common/utils/StarTotal'
+import ExpandableSection from './ExpandableSection'
+import DependencyTable from './DependencyTable'
 
-const NpmSection = ({ project }) => (
-  <section className="inner npm-section">
-    <a
-      href={`https://www.npmjs.com/package/${project.npm}`}
-      style={{ display: 'flex', alignItems: 'center', marginBottom: '.5rem' }}
-      target="_blank"
-    >
-      <img
-        src="/logos/npm.svg"
-        alt="NPM"
-        className="npm"
-        height="7"
-        width="18"
-        style={{ marginRight: '.25rem' }}
-      />
-      <span className="link" style={{ marginRight: '.25rem' }}>
-        {project.npm}
-      </span>
-      <span className="version text-secondary">{project.version}</span>
-    </a>
-    <Dependencies project={project} />
-  </section>
-)
+const DependenciesContainer = styled.div`
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+  .inline-list > *:not(:last-child) {
+    margin-right: 0.5rem;
+  }
+`
 
 const Dependencies = ({ project }) => {
-  const count = project.dependency_count
+  const { dependencies } = project.npm
+  if (!dependencies) return <span>Loading dependencies...</span>
+  const count = dependencies.length
   if (count === 0) return <span>No dependencies</span>
-  return project.dependencies ? (
-    <DependencyList dependencies={project.dependencies} />
-  ) : (
-    <span>Loading...</span>
-  )
+  return <DependencyList dependencies={dependencies} />
 }
 
 const DependencyList = ({ dependencies }) => (
   <Toggle>
     {({ on, getTogglerProps }) => (
-      <div className="dependencies">
-        <a className="toggler" {...getTogglerProps()}>
-          <span
-            className={`octicon octicon-triangle-${on ? 'down' : 'right'} icon`}
-          />{' '}
+      <DependenciesContainer>
+        <ExpandableSection on={on} getTogglerProps={getTogglerProps}>
           {dependencies.length}
           {' dependencies'}
-        </a>
+        </ExpandableSection>
         {!on && <DependencyListPreview dependencies={dependencies} />}
         <div>
           {on && (
@@ -59,7 +43,7 @@ const DependencyList = ({ dependencies }) => (
             />
           )}
         </div>
-      </div>
+      </DependenciesContainer>
     )}
   </Toggle>
 )
@@ -67,14 +51,16 @@ const DependencyList = ({ dependencies }) => (
 const DependencyListPreview = ({ dependencies }) => (
   <span className="inline-list" style={{ marginLeft: '.5rem' }}>
     {dependencies.map(packageName => (
-      <span key={packageName}>{packageName}</span>
+      <span className="text-secondary" key={packageName}>
+        {packageName}
+      </span>
     ))}
   </span>
 )
 
 const DependencyFullList = ({ packages }) => {
   return (
-    <table className="block-list">
+    <DependencyTable className="block-list">
       <thead>
         <tr>
           <td>Package on npm</td>
@@ -94,9 +80,6 @@ const DependencyFullList = ({ packages }) => {
                 <span>
                   <Link
                     to={`/projects/${npmPackage.project.slug}`}
-                    styleX={{
-                      textDecoration: 'underline'
-                    }}
                     className="text-secondary"
                   >
                     {npmPackage.project.name}{' '}
@@ -111,10 +94,10 @@ const DependencyFullList = ({ packages }) => {
           </tr>
         ))}
       </tbody>
-    </table>
+    </DependencyTable>
   )
 }
 
 const ConnectedDependencyFullList = withPackageData(DependencyFullList)
 
-export default NpmSection
+export default Dependencies