+++ src/components/ProjectView/GithubTab/NpmCard/FileSize.js
@@ -1,0 +1,8 @@
+import prettyBytes from '../../../../helpers/pretty-bytes'
+
+const FileSize = ({ value }) => {
+  if (isNaN(value)) return null
+  return prettyBytes(value)
+}
+
+export default FileSize