+++ src/components/ProjectView/GithubTab/index.js
@@ -1,11 +1,12 @@
 import React from 'react'
 
 import Tabs from '../Tabs'
-import Readme from './Readme'
 import Header from './Header'
 import ProjectHeader from '../ProjectHeader'
-import Trends from './Trends'
 import ProjectTabsContent from '../ProjectTabsContent'
+import NpmCard from './NpmCard'
+import TrendsCard from './TrendsCard'
+import ReadmeCard from './ReadmeCard'
 
 const GitHub = props => {
   const { project } = props
@@ -16,8 +17,9 @@ const GitHub = props => {
       <ProjectTabsContent>
         <Header {...props} />
       </ProjectTabsContent>
-      <Trends {...props} />
-      <Readme {...props} />
+      {project.packageName && <NpmCard {...props} />}
+      <TrendsCard {...props} />
+      <ReadmeCard {...props} />
     </div>
   )
 }