+++ src/components/common/Card.js
@@ -9,18 +9,6 @@ const Card = styled.div`
   background-color: #fff;
   vertical-align: top;
   border: 1px solid ${cardBorderColor};
-  .header {
-    padding: 0.5rem 1rem;
-    font-size: 1em;
-    border-bottom: 1px solid #ccc;
-  }
-  .header .counter,
-  .header .comment {
-    color: rgba(255, 255, 255, 0.7);
-  }
-  .header .big {
-    font-size: 1.5rem;
-  }
   .inner {
     padding: 1rem;
   }
@@ -39,11 +27,32 @@ const Card = styled.div`
   }
 `
 
+Card.Header = styled.div`
+  padding: 0.5rem 1rem;
+  font-size: 1em;
+  border-bottom: 1px solid ${cardBorderColor};
+  .counter,
+  .comment {
+    color: rgba(255, 255, 255, 0.7);
+  }
+  .big {
+    font-size: 1.5rem;
+  }
+`
+
+Card.Body = styled.div``
+
 Card.Section = styled.div`
   padding: 1em;
   &:not(:first-child) {
     border-top: 1px dashed ${cardBorderColor};
   }
+`
+
+Card.Footer = styled.div`
+  padding: 0.5rem 1rem;
+  text-align: center;
+  border-top: 1px solid ${cardBorderColor};
 `
 
 export default Card